Stop exporting internal class CollabWindowTracker from the container-loader package
